Fix bug with setting options on LocationManager
LocationProvider is not constrained to be a class, so when syncOptions()
was called, it triggered the setter for the property that stored it
rather than merely setting a property on the object itself. This caused
the location provider to be stopped and restarted and have its delegate
reconfigured each time options were set on LocationManager.