Trying to delegate panning exclusively to pan gesture recogniser
The problem is that when map starts being dragged, then it being zoomed in/out(before drag ends) - consecutive drags are being applied relative to the original zoom level - e.g. when the map is zoomed out(relative to the initial zoom level) - it barely moves, vice versa for zoomed in(map flies).