docs: Update support and community contact informationi in README (#1592)
Updated the contact information in the README file to provide users with more specific options for getting support and connecting with the developer community. Added a mention of our Discord channel. This update aims to encourage users to utilize the appropriate support channels instead of using the issue tracker for questions.