Support custom map resizing (#1948)
The default behavior is jumpy when animation is reducing view sizes – MTKView goes straight to the final bounds sizes as soon as the values are assigned to the property. 
However, MapKit has a completely different approach – when the new size is smaller, the drawableSize resizing operation is delayed until the animation is complete. The challenge here is that we have no access to the completion handler of animating code.
We also cannot make a redraw on each step of animation since that would cause a lot of CPU and GPU work on the renderer and frames would arrive too late to display them properly.

### Solution overview

The solution consists of two components:
1. Detect when we need to resize drawableSize
2. Calculate expected drawableSize values.

The correct moment for resizing varies by the direction of the animation. If the animation is going to increase the size of the MapView at least in one dimension when we have to update the drawable size immediately. So MTKView would cover new portions of the superview with an extended map. 
And if animation is going to reduce MapView size, we must hold the drawable size update until the animation is complete to prevent showing superview content instead of the map view. 

Because of that, the calculation of size differs at the start and end of the animation. In the end, we always use the final size that the animation was driving to. When animation begins, we have to use maximum values of what we have and what we are going to have. That trick guarantees that we will always cover superview content with MTKView.

The most challenging part however was the detection of animation completion. Since we have no access to the animation calling API and associated completion handler, the best approach I've found is to use the `CATransaction/setCompletionHandler` API. Implementation abuses the fact that new bounds size value is assigned in the same [CATransaction](https://developer.apple.com/documentation/quartzcore/catransaction) where animation is created. 

Another problem is to keep non-animating size updates as instant as it was and it is possible that bounds size would be updated without animation but there would be an active unrelated CATransaction with a pretty long duration. To address that issue the `add(CAAnimation)` method checks the `keyPath` value of animation (or nested animations if it is a `CAAnimationGroup`) for the start of the `bounds` string. It is not possible to check strictly for `bounds` values because CoreAnimation supports KVC and keys like `bounds.size` or `bounds.size.height` as also valid for size animations. 

#### Constraints 

Previously, the `metalView` constraints were aligned to the MapView bounds with mapboxMap size synchronized on the `layoutSubviews` calls. That approach doesn't work anymore because the size of the `metalView` can be bigger than the MapView (in case of animating to the bigger size since we need to cover new portions of the view preliminary).
Also now we need immediate control over the `metalView` size to avoid glitches when a new rendered map is displayed in outdated bounds size. To address that issue I got rid of the constraints usage for `metalView` and replaced it with manual `bounds.size` updates along with the `drawableSize` updates. That is also a great place to synchronize `mapboxMap` sizes. By the way, it is still unclear why GL-Native needs both `drawableSize` and `mapboxMap.size` that are kinda the same but we can assume that this requirement comes from the cross-platform support with Metal arrived later.
Another reason to let constraints go is that updating constraint values for width and size would lead to the bounds.size update on the next run loop cycle which mean that we would have a glitch with the new rendered map texture displayed in the wrong size map view for a 1 frame.

### Alternative implementations

#### CAAnimationDelegate 

Another promising idea was to overwrite the add animation method and assign a custom [CAAnimationDelegate](https://developer.apple.com/documentation/quartzcore/caanimationdelegate) with proxying capabilities to the original one. That gives us access to the `animationStart` and `animationStop` methods which are the best timings to recalculate drawableSize. While the solution was great for an UIView.animate set of methods, it didn't work with the `UIViewPropertyAnimator`. UIView.animate methods remove associated CAAnimation at the end while the latter doesn't. Next time animation happens it looks for an existing animation and reuses it. However, as soon as we change the delegate value `UIViewPropertyAnimator` won't recognize the animation and will create a new copy instead. Unfortunately, it is not possible to set the original delegate at the end of the animation as it causes a runtime exception (also the instance of CAAnimation is a new one).

I still suppose that we can improve the current implementation based on that idea but be smarter like copying the animation (it supports NSCopying protocol) and applying it to the custom value on the layer.

#### DisplayLink and presentation layer 

My first implementation was founded on the idea that the layer has [`model`](https://developer.apple.com/documentation/quartzcore/calayer/1410853-model) and [`presentations`](https://developer.apple.com/documentation/quartzcore/calayer/1410744-presentation) representations. The latter has access to approximate render tree data that can represent the actual values on the screen. 
The basic idea was to detect when the presentation bounds size is not equal to the model one and compare it to the caches bounds size value of the model. This is how we could detect animations and postpone `drawableSize` update until presentation value would reach the end value. 
Pros:
- the solution is simple and as it just observes actual changes it doesn't need to support any specific types of animations
Cons:
- extra workload is added to each display link trigger meaning that we will spend more CPU cycles each frame even when resizing happens
- we do pause display link when application loses a focus and that might cause glitches when user returns to the app that was in the middle of the animation
- there were some issue with a device rotation animation