Fix glitch on chained camera animations (#1970)
BasicCameraAnimator utilizes CoreAnimation to interpolate data in the animation. We have a special class CameraView where some default _animatable_ layer properties (like `borderWidth` or `cornerRadius`) are used like `zoomLevel` or `bearing`. On each CADisplayLink update, we later read the actual data from `presentation()` layer and reconstruct interpolated CameraOptions. 
The issue was here because the `presentation()` layer represents value at the start of the current transaction. In chained animations, the first frame of the following animations would be used both for assigning starting values to the CameraView and updating the interpolated data. However, the `presentation()` layer would have initial `CALayer()` information and so on wrong data would be populated to the CameraOptions.
This is not a regression but this issue was never observed in Maps v10 version because the starting position of the camera did use bearing and pitch as `0` values (which is the same as empty CALayer() reconstruction). Do address that I've set all animated CameraOptions values to non-zero values.