Reduce amount of iterations in geojson update integration test (#3286)
Decrease amount of iterations from 90 to 3(one run for each add, update,
delete operations) for GeoJSON update integration test.

Historically this test also doubled as a benchmark hence the number of
iterations, but since we don't use that data it's safe to reduce the
number of iterations.

cc @mapbox/sdk-ci

GitOrigin-RevId: fd9e14bd6ff9351316cd3005d5db546ba603b8de