Fix info button not having correct accessibility label (#3502)
Because `InfoButtonOrnament` is a `UIView` the `accesibilityLabel`
property is ignored. To address this we need to explicitly mark it as an
accessibility element and assign proper trait to it.
Alternative solution could be to assign accessibility label to the
internal `UIButton` instance, but I've decided against it as then the
label won't be configurable by external customers.

Addresses: https://mapbox.atlassian.net/browse/MAPSIOS-1825

cc @mapbox/sdk-ci

GitOrigin-RevId: 7e32856f09c39a86bdbc45a7b5acfe174dd426d9